bot/requirements/base.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
emoji==2.14.0
fastapi==0.115.5
frozenlist==1.5.0
gunicorn==20.1.0
h11==0.14.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.1.3
pdfkit==1.0.0
pillow==11.0.0
prettytable==3.12.0
prometheus-async==22.2.0
prometheus_client==0.21.0
propcache==0.2.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.18
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
wrapt==1.17.0
WTForms==3.1.2
yarl==1.18.0
